fix(theme): mark theme package exports as built to prevent double CSS injection
Theme packages ship pre-compiled CSS via `xds theme build`, but the
exported theme objects lacked the `__built` flag. When consumers imported
both the theme object and the CSS file, XDSTheme would inject a duplicate
<style> block at runtime — applying every override twice.

Changes:
- Add `markBuilt()` utility to @xds/core/theme
- Each theme package's index.ts now wraps its export with markBuilt()
- Extract brutalist theme definition to its own file (consistency)
- Add unit tests for markBuilt()
