+++ packages/themes/daily/src/index.ts
@@ -1,2 +1,5 @@
-export {dailyTheme} from './dailyTheme';
+import {markBuilt} from '@xds/core/theme';
+import {dailyTheme as _dailyTheme} from './dailyTheme';
+
+export const dailyTheme = markBuilt(_dailyTheme);
 export {dailyIconRegistry} from './icons';
