+++ packages/themes/default/src/index.ts
@@ -1,2 +1,5 @@
-export {defaultTheme} from './defaultTheme';
+import {markBuilt} from '@xds/core/theme';
+import {defaultTheme as _defaultTheme} from './defaultTheme';
+
+export const defaultTheme = markBuilt(_defaultTheme);
 export {defaultIconRegistry} from './icons';
