+++ packages/themes/meta/src/index.ts
@@ -1,2 +1,5 @@
-export {metaTheme} from './metaTheme';
+import {markBuilt} from '@xds/core/theme';
+import {metaTheme as _metaTheme} from './metaTheme';
+
+export const metaTheme = markBuilt(_metaTheme);
 export {metaIconRegistry} from './icons';
