+++ packages/themes/neutral/src/index.ts
@@ -1,2 +1,5 @@
-export {neutralTheme} from './neutralTheme';
+import {markBuilt} from '@xds/core/theme';
+import {neutralTheme as _neutralTheme} from './neutralTheme';
+
+export const neutralTheme = markBuilt(_neutralTheme);
 export {neutralIconRegistry} from './icons';
