+++ packages/themes/whatsapp/src/index.ts
@@ -1,2 +1,5 @@
-export {whatsappTheme} from './whatsappTheme';
+import {markBuilt} from '@xds/core/theme';
+import {whatsappTheme as _whatsappTheme} from './whatsappTheme';
+
+export const whatsappTheme = markBuilt(_whatsappTheme);
 export {whatsappIconRegistry} from './icons';
